requirements.txt
flask
flask-cors
openai
anthropic
stripe
python-dotenv
requests
resend
gunicorn
